requirements.txt
psycopg2-binary
pyyaml
pandas
requests
numpy
matplotlib